requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
requests>=2.31.0
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
# eventlet>=0.33.0  # Removed due to Python 3.13 compatibility issues
pyaudio>=0.2.11
wave>=0.0.2
openai>=1.0.0
python-dateutil>=2.8.0
google-generativeai>=0.3.0
diffusers>=0.25.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.2
Pillow>=10.2.0
